+++ control/pscopehat/planktoscope/imagernew/__init__.py
@@ -1,0 +1,1 @@
+"""imagernew provides high-level functionality for performing image acquisition."""